Separated the Login and Logout unit tests into their own test files to modularize testing. The only test added is LoginTest.test_unique_tokens_for_different_users to ensure different users will be assigned unique access tokens
